tests/wpt/web-platform-tests/css/css-align/parsing/column-gap-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Box Alignment Level 3: parsing column-gap with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-align-3/#propdef-column-gap">
<meta name="assert" content="column-gap supports the full grammar '<length-percentage> | normal'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("column-gap", "normal");

test_valid_value("column-gap", "0", "0px");
test_valid_value("column-gap", "1px");
test_valid_value("column-gap", "calc(2em + 3ex)");
test_valid_value("column-gap", "4%");
test_valid_value("column-gap", "5vmin");
</script>
</body>
</html>
